Introduce two new docker containers
First one is codenvy/che-test allowing to check a local or remote Eclipse Che instance by creating a workspace, starting it and displaying console output
The other one is codenvy/che-file allowing to boot a local Eclipse Che by using a local directory. It will mount in Che this local folder as being a project.

Change-Id: Ie794b41f89608f86e589f1aeea06d63f0d9f2edc
Signed-off-by: Florent BENOIT <fbenoit@codenvy.com>