Fix OpenShiftConnector.inspectImage() to resolve issue in OS 3.6+
OpenShiftConnector.inspectImage() uses pods to figure out a workspaces
docker image. Previously, this was not a problem as it would match
the imagestream tag we create earlier. However, in OpenShift 3.6 and
higher, it seems like pods created by deployments use the sha256 of
the image as an identifier, and so this fails. This means that
workspaces fail to start due to imagestream not found.

To resolve this issue, we instead get the image name through the
deployment, which does have the correct value.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>