CHE-8557: No Dto available for FormattingOptions (#8784)
* CHE-8557: No Dto available for FormattingOptions

The Dto isn't needed for FormattingOptions as it is really a specialized
Map and the types that contain a FormattingOptions field handle
the field as a Map during JSON serialize/deserialize

Signed-off-by: Jonah Graham <jonah@kichwacoders.com>

* CHE-8557: Test cases for Either deserializing

Signed-off-by: Jonah Graham <jonah@kichwacoders.com>

* CHE-8557: Fix Either handling when deserializing on the server

Signed-off-by: Jonah Graham <jonah@kichwacoders.com>

* CHE-8557: Use actual type for Map conversions

The is necessary to ensure handling types which are of type Map are
still instantiated in their real type. Consider
DocumentFormattingParams.setOptions() which takes a FormattingOptions
class.

Signed-off-by: Jonah Graham <jonah@kichwacoders.com>

* CHE-8557: Register LSP4J's Either adapter factory

This is effectively a follow up CHE-3103 which uses Gson directly
to serialize/deserialize Json. To support LSP4J's Either types,
the either type adapter factory is needed.

Signed-off-by: Jonah Graham <jonah@kichwacoders.com>

* load EitherTypeAdapterFactory for registration

Signed-off-by: Valeriy Svydenko <vsvydenk@redhat.com>