Wait for children nodes to be loaded on refresh project explorer after adding new resource (#9470)
Expand package node in Project Explorer if it contains child node.
We have TreeResourceRevealer that handle's nodes loading, witch can expand such nodes, but it catches event with children nodes that received from local storage. It means that children can be not actual and added package can be not expanded in this case.
Added a delay for disabling handler.