Refactor external server exposer strategy to expose service port one by one
Previously expose server strategy had to match servers with services port,
but sometimes it is impossible like in case of JWT proxy. Because in this case,
server config contains port that is different to service port. Like `wsagent`
server is exposed on 4401 original service port, but it will be publicly
accessible via 4911 (random one) JWTProxy service port.