Fix start of a workspace with Theia on k8s infra
When default-host strategy for server links evaluation is used
on k8s infra Theia server has a link without trailing slash.
This leads to a situation when Theia doesn't load because its
main javascript is not loaded properly (URL gets truncated to
parent and then name of the script is added).
This PR adds a trailing slash to Theia server definition to
fix this.
Theia stack for Openshift recice is changed to align
configuration of dockerimage and openshift stacks.
Signed-off-by: Oleksandr Garagatyi <ogaragat@redhat.com>