Fix sequential start of workspace machines
Fix an issue where `KubernetesDeployments#deploy` blocks on launching a
workspace until the underlying Deployment is 'ready', which is taken to
mean when all pods in the deployment are 'running'. This causes
various issues that rely on watching a Pod before it's fully running.

With this change, the method returns as soon as there is a Pod to
return.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>