Ensure plugin broker works with ephemeral workspaces
Changes necessary to ensure that plugin broker can work with ephemeral
workspaces, which, since they use emptyDir volumes, cannot share storage
between pods (as in the separate pod that runs plugin broker). This is
done by adding the plugin broker as an init container on ephemeral
workspaces.

Details:
- Change EphemeralWorkspaceAdapter to make check methods static, add
method to "make" a workspace ephemeral (to ensure how we check for
ephemeral workspaces is done in one place)
- Make PVC strategies include init containers in assigning volumes, to
ensure broker's volumes are correctly set when running as an init
container
- Add KubernetesBrokerInitContainerApplier, which modifies the workspace
env to include the broker as an init container.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>