Inject pods into other deployments (#15890)
Introduce a concept of injectable pod that provisioners can add to the environment. Such pods are merged into the pods that require them.

Use the injectable pods as a new way of deploying JWT proxy that becomes part of the workspace pod.

Secure servers are no longer exposed using services and JWT proxy proxying to those services. Rather, the JWT proxy assumes the secure servers listen on localhost and proxies the localhost on given port directly, taking advantage of the fact that it is inside the workspace pod.

Signed-off-by: Lukas Krejci <lkrejci@redhat.com>